business_manager/src/requirements.txt
amqp==1.4.9
anyjson==0.3.3
arrow==0.8.0
billiard==3.3.0.23
celery==3.1.23
click==6.6
configobj==5.0.6
Django==1.7.7
django-common-helpers==0.8.0
django-cors-headers==1.1.0
django-cron==0.4.6
django-gearman-commands==0.6.1
gearman==2.0.2
kombu==3.0.35
mycli==1.7.1
MySQL-python==1.2.5
numpy==1.11.1
pony==0.6.5
powerline-status==2.4
prettytable==0.7.2
prompt-toolkit==1.0.3
pycrypto==2.6.1
pyExcelerator==0.6.4.1
Pygments==2.1.3
pymongo==3.2.2
PyMySQL==0.7.4
python-dateutil==2.5.3
pytz==2016.4
redis==2.10.5
requests==2.10.0
simplejson==3.8.2
six==1.10.0
sqlparse==0.1.19
thrift==0.9.3
wcwidth==0.1.6
xlrd==1.0.0
xlwt==1.1.2
celery==3.1.24
django-celery==3.1.17
djangorestframework==3.3.3
openpyxl==2.4.0
